Update patch set 2
Patch Set 2:

According to the original commit, it's "microlags", so maybe try slowing down the animation speed using the developer options, and compare flicking around home screens with lots of widgets or long lists with lots of pictures, in builds with and without it.

I'm currently working on moving to a new drive, so I haven't gotten a fresh build in a while, but I'll be A/B testing this change as soon as that happens.

Patch-set: 2