Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I would preffer to left the code as google did. Is their work and unless we really find those warning can produce any benefit to be removed is not a good idea to move from theirwork without knowing why they left the code as it is. Linux kernel is full of warnings but some of them are intentional from the coder. Is not bad to have warning on the code if you understant that the final compilation result is going to be ok on your device.

Patch-set: 1