Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I fixed these warnings because they irritated me, not because of any issues related to them. I fixed them for my own personal kernel because I can. There was no technical reason behind it.

None of them are intentional warnings. It's sloppy coding by Samsung which has been largely ignored by Google. It reeks of the 'It builds, it boots, it works, stuff it' attitude.

I don't think there is any point in merging something which isn't understood by the person who wants to merge it though, so for that reason I'm going to -1 it.

Patch-set: 1