Update patch set 1
Patch Set 1:

I don't try to be rude with your work Will (I appreciate it). Jut I try to follow allways the same rule, just change what is a bennefit not other things, since Google maintains the Kernel on the new versions and that way is aesyer to follow their work.

Rafael

Patch-set: 1