Update patch set 1
Patch Set 1: (1 inline comment)

Looks good. Keeps crespo on parity with similar devices.

I know jt1134 did the work squashing them (for aries IIRC), but Todd Poynor actually seems to be the author on the majority of commits, maybe we should give Todd authorship.

I'm building now, and will report back, but these patches are already in aries and d2, so it should be fine.

Patch-set: 1