Update patch set 1
Patch Set 1:

I'm not sure what it is about this change, but it makes it take around 3 seconds for the screen to turn on when I have it. It doesn't seem to me like any of the code should be run at screen on, but I'm not sure what else it could be.

Patch-set: 1