Update patch set 1
Patch Set 1:

It shows up for me as a CDROM in Windows, but my laptop won't boot off of it in CDROM mode. Do you know if that could just be an issue with my laptop, or if it's something more? I can try testing on some more computers soon.

I was trying to boot off the debian 7.1 i386 net install iso, so that should work...

Patch-set: 1